🚧 ajustando os icones de enfrentamento
